[DS-1920] Organized new models
